pyproject.toml
[project]
name = "phonezilla"
version = "0.1.0"
description = "A Django-based phone inventory management system"
authors = [
    { name = "Gahan Saraiya", email = "gahan@example.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "backcall==0.1.0",
    "certifi>=2024.7.4",
    "chardet>=5.2.0",
    "colorama>=0.4.6",
    "decorator>=5.1.1",
    "django>=6.0.1",
    "django-filter>=25.2",
    "django-jet-reboot>=1.3.10",
    "django-money>=3.5.4",
    "django-nested-admin>=4.1.6",
    "django-nested-inline>=0.4.0",
    "django-phonenumber-field>=8.4.0",
    "djangorestframework>=3.16.1",
    "idna>=3.7",
    "ipython>=8.26.0",
    "ipython-genutils==0.2.0",
    "itypes==1.1.0",
    "jedi>=0.19.1",
    "jinja2>=3.1.4",
    "markdown>=3.6",
    "markupsafe>=2.1.5",
    "mysqlclient>=2.2.4",
    "parso>=0.8.4",
    "pexpect>=4.9.0",
    "phonenumbers>=9.0.21",
    "pickleshare==0.7.4",
    "pillow>=10.4.0",
    "prompt-toolkit>=3.0.47",
    "ptyprocess==0.5.2",
    "py-moneyed>=3.0",
    "pygments>=2.18.0",
    "python-monkey-business==1.0.0",
    "pytz>=2024.1",
    "reportlab>=4.2.2",
    "requests>=2.32.3",
    "selenium>=4.23.1",
    "simplegeneric==0.8.1",
    "six>=1.16.0",
    "traitlets>=5.14.3",
    "uritemplate>=4.1.1",
    "urllib3>=2.2.2",
    "wcwidth>=0.2.13",
]

[project.scripts]
phonezilla = "manage:main"

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "ruff>=0.14.10",
]


